+++ src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/SemanticTokens/RemoteCSharpSemanticTokensProvider.cs
@@ -4,16 +4,13 @@
 using System;
 using System.Collections.Immutable;
 using System.Composition;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Razor;
 using Microsoft.AspNetCore.Razor.Telemetry;
 using Microsoft.CodeAnalysis.ExternalAccess.Razor.Cohost.Handlers;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 using Microsoft.CodeAnalysis.Razor.SemanticTokens;
 using Microsoft.CodeAnalysis.Razor.Workspaces;
-using Microsoft.CodeAnalysis.Remote.Razor.ProjectSystem;
 using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.CodeAnalysis.Remote.Razor.SemanticTokens;
@@ -30,12 +27,7 @@ internal class RemoteCSharpSemanticTokensProvider(IFilePathService filePathServi
         using var _ = _telemetryReporter.TrackLspRequest(nameof(SemanticTokensRange.GetSemanticTokensAsync), Constants.ExternalAccessServerName, correlationId);
 
         // We have a razor document, lets find the generated C# document
-        var generatedDocument = GetGeneratedDocument(documentContext);
-
-        var csharpSourceText = await documentContext.GetCSharpSourceTextAsync(cancellationToken).ConfigureAwait(false);
-
-        // HACK: We're not in the same solution fork as the LSP server that provides content for this document
-        generatedDocument = generatedDocument.WithText(csharpSourceText);
+        var generatedDocument = await documentContext.GetGeneratedDocumentAsync(_filePathService, cancellationToken).ConfigureAwait(false);
 
         var data = await SemanticTokensRange.GetSemanticTokensAsync(
             generatedDocument,
@@ -44,20 +36,5 @@ internal class RemoteCSharpSemanticTokensProvider(IFilePathService filePathServi
             cancellationToken).ConfigureAwait(false);
 
         return data;
-    }
-
-    private Document GetGeneratedDocument(VersionedDocumentContext documentContext)
-    {
-        var snapshot = (RemoteDocumentSnapshot)documentContext.Snapshot;
-        var razorDocument = snapshot.TextDocument;
-        var solution = razorDocument.Project.Solution;
-
-        // TODO: A real implementation needs to get the SourceGeneratedDocument from the solution
-
-        var projectKey = razorDocument.Project.ToProjectKey();
-        var generatedFilePath = _filePathService.GetRazorCSharpFilePath(projectKey, razorDocument.FilePath.AssumeNotNull());
-        var generatedDocumentId = solution.GetDocumentIdsWithFilePath(generatedFilePath).First(d => d.ProjectId == razorDocument.Project.Id);
-        var generatedDocument = solution.GetDocument(generatedDocumentId).AssumeNotNull();
-        return generatedDocument;
     }
 }
